+++ modules/chain/dap_chain_ch_pkt.c
@@ -31,7 +31,7 @@ static void s_chain_pkt_fill(dap_chain_ch_pkt_t *a_pkt, dap_chain_net_id_t a_net
             .hdr = { .version = a_version,
                      .data_size = a_version == DAP_CHAIN_CH_PKT_VERSION_LEGACY
                                         ? 0
-                                        : a_data_size,
+                                        : (uint32_t)a_data_size,
                      .net_id = a_net_id,
                      .cell_id = a_cell_id,
                      .chain_id = a_chain_id }
