+++ modules/common/dap_chain_common.c
@@ -127,7 +127,7 @@ size_t dap_chain_addr_from_str_array(const char *a_addr_str, dap_chain_addr_t **
     dap_return_val_if_pass(!a_addr_str || !a_addr, 0);
     size_t l_count = dap_str_symbol_count(a_addr_str, ',') + 1;
     dap_chain_addr_t *l_addr = DAP_NEW_Z_COUNT_RET_VAL_IF_FAIL(dap_chain_addr_t, l_count, 0);
-    char **l_addr_str_array = dap_strsplit(a_addr_str, ",", l_count);
+    char **l_addr_str_array = dap_strsplit(a_addr_str, ",", (int)l_count);
     if (!l_addr_str_array) {
         DAP_DELETE(l_addr);
         return 0;
