+++ modules/mempool/dap_chain_mempool.c
@@ -335,7 +335,7 @@ char *dap_chain_mempool_tx_create(dap_chain_t *a_chain, dap_enc_key_t *a_key_fro
 
     }
     if (a_tx_num > 1) {
-        uint32_t l_tx_num = a_tx_num;
+        uint32_t l_tx_num = (uint32_t)a_tx_num;
         dap_chain_tx_tsd_t *l_out_count = dap_chain_datum_tx_item_tsd_create(&l_tx_num, DAP_CHAIN_DATUM_TRANSFER_TSD_TYPE_OUT_COUNT, sizeof(uint32_t));
         dap_chain_datum_tx_add_item(&l_tx, l_out_count);
         DAP_DELETE(l_out_count);
@@ -633,7 +633,7 @@ char *dap_chain_mempool_tx_create_extended(dap_chain_t *a_chain, dap_enc_key_t *
 
     }
     if (a_tx_num > 1) {
-        uint32_t l_tx_num = a_tx_num;
+        uint32_t l_tx_num = (uint32_t)a_tx_num;
         dap_chain_tx_tsd_t *l_out_count = dap_chain_datum_tx_item_tsd_create(&l_tx_num, DAP_CHAIN_DATUM_TRANSFER_TSD_TYPE_OUT_COUNT, sizeof(uint32_t));
         dap_chain_datum_tx_add_item(&l_tx, l_out_count);
         DAP_DELETE(l_out_count);
